+++ Models/TemplateModel.cs
@@ -1,9 +1,14 @@
+using YamlDotNet.Serialization;
+
 namespace etvctl.Models;
 
 public class TemplateModel
 {
-    //public List<ChannelModel> Channels { get; set; } = [];
+    [YamlMember(Alias = "ffmpeg_profiles")]
+    public List<FFmpegProfileModel> FFmpegProfiles { get; set; } = [];
+
+    [YamlMember(Alias = "smart_collections")]
     public List<SmartCollectionModel> SmartCollections { get; set; } = [];
 
-    public bool IsEmpty() => SmartCollections.Count == 0;
+    public bool IsEmpty() => FFmpegProfiles.Count == 0 && SmartCollections.Count == 0;
 }