+++ Planning/YamlWriter.cs
@@ -13,7 +13,7 @@ public static class YamlWriter
         CancellationToken cancellationToken = default)
     {
         var serializer = new StaticSerializerBuilder(new YamlStaticContext())
-            .ConfigureDefaultValuesHandling(DefaultValuesHandling.OmitNull)
+            .ConfigureDefaultValuesHandling(DefaultValuesHandling.OmitNull | DefaultValuesHandling.OmitEmptyCollections)
             .WithNamingConvention(UnderscoredNamingConvention.Instance)
             .WithTypeConverter(new FileOrganizationTypeConverter())
             .Build();
@@ -23,6 +23,76 @@ public static class YamlWriter
         // write each resource type
 
         string singleFileName = Path.Combine(config.Template!, "ersatztv.yml");
+
+        // ffmpeg profiles
+        string ffmpegProfilesFolderName = Path.Combine(config.Template!, "ffmpeg_profiles");
+        string ffmpegProfilesFileName = Path.Combine(config.Template!, "ffmpeg_profiles.yml");
+
+        var ffmpegProfilesOrganization = config.Organization?.Overrides?.FFmpegProfile ?? defaultOrganization;
+        if (ffmpegProfilesOrganization is FileOrganization.FilePerType)
+        {
+            if (Directory.Exists(ffmpegProfilesFolderName))
+            {
+                Directory.Delete(ffmpegProfilesFolderName, true);
+            }
+
+            await File.WriteAllTextAsync(
+                ffmpegProfilesFileName,
+                serializer.Serialize(templateModel.FFmpegProfiles),
+                cancellationToken);
+
+            // dont writing ffmpeg profiles; don't write again
+            templateModel.FFmpegProfiles.Clear();
+        }
+        else if (ffmpegProfilesOrganization is FileOrganization.FilePerResource)
+        {
+            if (!Directory.Exists(ffmpegProfilesFolderName))
+            {
+                Directory.CreateDirectory(ffmpegProfilesFolderName);
+            }
+            else
+            {
+                Directory.Delete(ffmpegProfilesFolderName, true);
+                Directory.CreateDirectory(ffmpegProfilesFolderName);
+            }
+
+            if (File.Exists(ffmpegProfilesFileName))
+            {
+                File.Delete(ffmpegProfilesFileName);
+            }
+
+            foreach (var ffmpegProfile in templateModel.FFmpegProfiles)
+            {
+                if (string.IsNullOrWhiteSpace(ffmpegProfile.Name))
+                {
+                    continue;
+                }
+
+                string fileName = Path.Combine(
+                    ffmpegProfilesFolderName,
+                    $"{ffmpegProfile.Name!.Replace(" ", "-")}.yml");
+                await File.WriteAllTextAsync(
+                    fileName,
+                    serializer.Serialize(ffmpegProfile),
+                    cancellationToken);
+            }
+
+            // done writing ffmpeg profiles; don't write again
+            templateModel.FFmpegProfiles.Clear();
+        }
+        else
+        {
+            if (File.Exists(ffmpegProfilesFileName))
+            {
+                File.Delete(ffmpegProfilesFileName);
+            }
+
+            if (Directory.Exists(ffmpegProfilesFolderName))
+            {
+                Directory.Delete(ffmpegProfilesFolderName, true);
+            }
+        }
+
         string collectionsFolderName = Path.Combine(config.Template!, "collections");
         string smartCollectionsFileName = Path.Combine(collectionsFolderName, "smart.yml");
         string smartCollectionsFolderName = Path.Combine(config.Template!, "collections", "smart");
@@ -59,6 +129,11 @@ public static class YamlWriter
             {
                 Directory.Delete(smartCollectionsFolderName, true);
                 Directory.CreateDirectory(smartCollectionsFolderName);
+            }
+
+            if (File.Exists(smartCollectionsFileName))
+            {
+                File.Delete(smartCollectionsFileName);
             }
 
             foreach (var smartCollection in templateModel.SmartCollections)